# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documents",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transactions/insert": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Insert new transaction (mockup api)",
                "operationId": "eea7d27d831e444fb9dd687136ad1119",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transaction_type",
                                    "token_type",
                                    "status",
                                    "payment_type",
                                    "system_type",
                                    "pay_currency",
                                    "amount",
                                    "per_price",
                                    "total_price",
                                    "sender_id",
                                    "receiver_id",
                                    "sell_order_id"
                                ],
                                "properties": {
                                    "transaction_type": {
                                        "description": "value: buy,sell,transfer,receive",
                                        "type": "string",
                                        "example": "buy"
                                    },
                                    "token_type": {
                                        "description": "value: BTC,ETH,XRP,DOGE",
                                        "type": "string",
                                        "example": "XRP"
                                    },
                                    "status": {
                                        "description": "value: pending,processing,completed,failed",
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "payment_type": {
                                        "description": "value: qrcode,credit",
                                        "type": "string",
                                        "example": "credit"
                                    },
                                    "system_type": {
                                        "description": "value: internal,external",
                                        "type": "string",
                                        "example": "internal"
                                    },
                                    "pay_currency": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "per_price": {
                                        "type": "decimal",
                                        "example": 500
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "example": 1000
                                    },
                                    "sender_id": {
                                        "type": "string",
                                        "example": "abc527bd-79d0-4799-a883-9c7257990d3a"
                                    },
                                    "receiver_id": {
                                        "type": "string",
                                        "example": "3535a148-2431-48ea-a1bf-314e52a8fb56"
                                    },
                                    "sell_order_id": {
                                        "type": "string",
                                        "example": "5ca2d56d-79ad-41e3-889f-0cdc12c77014"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "9390cea0-6fa6-41f3-a003-b5c3fb69db3a"
                                                },
                                                "create_date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-05-16 11:35:59"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Insert transaction failed"
                    }
                }
            }
        },
        "/api/transactions/list": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Transactions List",
                "operationId": "04c19f98b51f99c74e79bee33700b513",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "member_id": {
                                        "type": "string",
                                        "example": "abc527bd-79d0-4799-a883-9c7257990d3a"
                                    },
                                    "keyword": {
                                        "description": "Search by transaction_id, receiver_name",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "sort": {
                                        "description": "Sort by last update_date => value: asc,desc",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "page": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "limit": {
                                        "type": "string",
                                        "example": "10"
                                    },
                                    "transaction_type": {
                                        "description": "value: buy,sell,transfer,receive",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "token_type": {
                                        "description": "value: BTC,ETH,XRP,DOGE",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "status": {
                                        "description": "value: pending,processing,completed,failed",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "payment_type": {
                                        "description": "value: qrcode,credit",
                                        "type": "string",
                                        "example": "qrcode,credit"
                                    },
                                    "system_type": {
                                        "description": "value: internal,external",
                                        "type": "string",
                                        "format": "email",
                                        "example": ""
                                    },
                                    "pay_currency": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "THB"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "string",
                                                    "example": "abc527bd-79d0-4799-a883-9c7257990d3\u0e34b"
                                                },
                                                "transaction_type": {
                                                    "type": "string",
                                                    "example": "buy"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "BTC"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "payment_type": {
                                                    "type": "string",
                                                    "example": "qrcode"
                                                },
                                                "system_type": {
                                                    "type": "string",
                                                    "example": "internal"
                                                },
                                                "pay_currency": {
                                                    "type": "string",
                                                    "example": "THB"
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": "2"
                                                },
                                                "per_price": {
                                                    "type": "decimal",
                                                    "example": "3457601.59"
                                                },
                                                "total_price": {
                                                    "type": "decimal",
                                                    "example": "6915203.18"
                                                },
                                                "sender_id": {
                                                    "type": "string",
                                                    "example": "abc527bd-79d0-4799-a883-9c7257990d3a"
                                                },
                                                "receiver_id": {
                                                    "type": "string",
                                                    "example": "abc527bd-79d0-4799-a883-9c7257990d3b"
                                                },
                                                "sell_order_id": {
                                                    "type": "string",
                                                    "example": "9390cea0-6fa6-41f3-a003-b5c3fb69db3a"
                                                },
                                                "create_date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-05-15 19:32:20"
                                                },
                                                "update_date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-05-15 19:32:20"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Not found data."
                    }
                }
            }
        }
    },
...
}